Fix NoStacktraceFormatter
This deals with a surprising gotcha in Python's logging system: the
result of formatException() is cached. If you want to ensure that a
custom formatException method is called, you need to explicitly clear
the exc_text attribute in format().

This is documented behavior, see https://bugs.python.org/issue29056
for background info.

In our setup, we had an earlier handler with a default formatter, so
NoStacktraceFormatter's method was entirely ignored.